src/Backend/requirements.txt
Flask==2.3.2
openai==0.28.1
python-dotenv==1.0.0
